Add a test function for moons
