+++ .github/workflows/pull-request.yml
@@ -10,6 +10,6 @@ jobs:
       - uses: actions/checkout@v4
       - uses: actions/setup-node@v3
         with:
-          node-version: "16"
+          node-version: "21"
       - run: npm install -g bats
       - run: bats -v
